Add files via upload
This Jupyter notebook contains a project I completed using a customer information database common in your everyday point-of-sale system found in retailers around the world. This particular dataset contains ten-thousand entries and has 13 unique pieces of customer information that can be used to determine spending habits, current occupation, the time of day a purchase was made and much more.